requirements.txt
aiofiles==24.1.0
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
distro==1.9.0
easydict==1.13
exceptiongroup==1.2.2
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.57.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy
onnx==1.17.0
onnxruntime==1.21.1
openai==1.76.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
prettytable==3.16.0
protobuf
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
deepface
tf-keras
torch
azure-storage-blob
